chnaged price from double to Int, sepearted search for products and vendor

+++ product/JwtDbApi/DTOs/ProductDto.cs
@@ -1,10 +1,12 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace JwtDbApi.DTOs
 {
     public class ProductDto
     {
         public string? ProdName { get; set; }
         public string? Description { get; set; }
-        public double Price { get; set; }
+        public int Price { get; set; }
         public string? ImageURL { get; set; }
         public DateTime? StartDate { get; set; }
         public int CategoryId { get; set; }
