+++ product/JwtDbApi/DTOs/ProductVendorDto.cs
@@ -1,8 +1,10 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace JwtDbApi.DTOs
 {
     public class ProductVendorDto
     {
-        public double Price { get; set; }
+        public int Price { get; set; }
         public int Quantity { get; set; }
         public int Visible { get; set; }
         public int ProductId { get; set; }
